+++ Plugins/ags_snowrain/ags_snowrain/ags_snowrain.cpp
@@ -810,7 +810,7 @@ void AGS_EngineShutdown()
   delete snow;
 }
 
-int AGS_EngineOnEvent(int event, int data)
+intptr_t AGS_EngineOnEvent(int event, intptr_t data)
 {
   if (event == AGSE_PREGUIDRAW)
   {
