+++ Plugins/agspalrender/agspalrender/agspalrender.h
@@ -7,7 +7,7 @@ namespace agspalrender
 {
   void AGS_EngineStartup(IAGSEngine *lpEngine);
   void AGS_EngineShutdown();
-  int AGS_EngineOnEvent(int event, int data);
+  intptr_t AGS_EngineOnEvent(int event, intptr_t data);
   int AGS_EngineDebugHook(const char *scriptName, int lineNum, int reserved);
   void AGS_EngineInitGfx(const char *driverID, void *data);
 }
