+++ Plugins/agstouch/agstouch/agstouch.cpp
@@ -76,7 +76,7 @@ void AGS_EngineShutdown()
 
 }
 
-int AGS_EngineOnEvent(int event, int data)
+intptr_t AGS_EngineOnEvent(int event, intptr_t data)
 {
   return 0;
 }
